Added missing folder and re-worked database population
